Removed Prefabs class

+++ Game/SceneInGame/HUD.cpp
@@ -56,12 +56,12 @@ void HUD::update() {
     }
     {   // HEALTH
         Rect rect = healthSprite_->getRect();
-        rect.x2 = (float)player_->getComponent<Health>()->getHealth();
+        rect.x2 = player_->getComponent<Health>()->getHealth();
         healthSprite_->setRect(rect);
     }
     {   // POWER
         Rect rect = powerSprite_->getRect();
-        rect.x2 = (float)player_->getComponent<Power>()->getPower();
+        rect.x2 = player_->getComponent<Power>()->getPower();
         powerSprite_->setRect(rect);
     }
     {   // ITEM SLOTS
